backend/requirements.txt
Flask
langchain
openai
python-dotenv
faiss-cpu
tiktoken
flask[async]
selenium